Move logic postprocess memory to utils > memories as a sub func
