+++ backend/requirements.txt
@@ -226,3 +226,4 @@ wcwidth==0.2.13
 websockets==12.0
 yarl==1.9.4
 pyogg @ git+https://github.com/TeamPyOgg/PyOgg@6871a4f
+opuslib==3.0.1
