+++ backend/routers/transcribe.py
@@ -15,7 +15,7 @@ from database.redis_db import get_user_speech_profile
 from models.memory import Memory, TranscriptSegment
 from models.message_event import NewMemoryCreated, MessageEvent, NewProcessingMemoryCreated
 from models.processing_memory import ProcessingMemory
-from utils.postprocessing import postprocess_memory_util
+from utils.memories.postprocess_memory import postprocess_memory as postprocess_memory_util
 from utils.audio import create_wav_from_bytes, merge_wav_files
 from utils.memories.process_memory import process_memory
 from utils.other.storage import upload_postprocessing_audio
@@ -216,7 +216,6 @@ async def _websocket_util(
             soniox_socket = await process_audio_soniox(
                 stream_transcript, speech_profile_stream_id, language, uid if include_speech_profile else None
             )
-
 
     except Exception as e:
         print(f"Initial processing error: {e}")
